Publish Releases In Background Job
When a release is published, but has too many changes, it can cause a
request timeout because it can't be fully published within the allotted
15 seconds in production. To prevent this, Workarea now runs all release
publishing in a background job. The success flash message for when a
release is published has been updated to inform users that changes may
take a little while to apply.

Fixes #1