Fix Deep Duplication of Swappable Lists
The `Workarea::SwappableList` class does not get duplicated correctly
when `Workarea.config.deep_dup` is used. This was observed while using
multi-site and attempting to change a swappable list for only one site.
Define the `#deep_dup` method to return a new object instead of referencing
the existing one.

ECOMMERCE-7038