Add Generic Activity Partials (#4)
Empty results were still being seen in the trash when a model that
doesn't explicitly have an activity partial defined is encountered. This
is due to the `render_activity_entry` helper rescuing an
`ActionView::TemplateError` to return a blank string. To resolve this
issue, models that are tracked by `Mongoid::AuditLog`, without an
explicit activity partial defined will be rendered using a generic
partial, showing the class name and ID of the audited model, as
something to render in the listing so that pages of blank results
aren't shown.