Display Referrer URL in tooltip on Order Attributes in Admin
Due to the length of URLs being displayed on Order Attributes in the
admin they will potentially break layout. Now they are displayed within
a tooltip behind a "View" link click. The resulting tooltip will prompt
the user to copy the contents of a text box containing the URL.

Fixes #60