Use private HTTP caching headers for responses with segmented content
If a page has segmented content, it can't be cached by any upstream HTTP
caches because the user's segments can change request-by-request.

Our solution is to use the headers we've been using for cached responses
if the page has no segmented content. If it does have segmented content,
change those headers to force refetching every time, while allowing the
server to return a 304 to eliminate sending unnecessary responses.

This is being done in a piece of middleware to ensure to
Rack::Cache the headers look the same. This allows us to still cache
complete responses in Rack::Cache for requests with segmented content.

This commit also refactors the middleware that sets this all up into a
single ApplicationMiddleware so it's easier to see everything going on
in one file.

WORKAREA-36