[DOCS] Rename and update doc for testing CC transactions
Rename "Test a Credit Card Transaction" to
"Manually Test Credit Card Transactions" and update the content.

The content was specific to a particular gateway and didn't make that
clear. These changes provide a generic solution in addition to the
specifics for the default gateway.

The title confused devs who reviewed this doc in a different PR,
thinking that it had to do with automated testing.

WORKAREA-13