Fix Incorrect Test Setup
The `Pricing::Calculators::Calculator.test_adjust` method accepts two
arguments, and expects the first argument is going to be of type
`Order`, but in a `TaxCalculator` test only a shipping was being passed
in. Update this test to use the correct syntax so that other downstream
projects that expect data to be on an Order won't get confused.