Ensure UTF-8 Encoding of Data File Import Samples
This ensures data file import samples are always treated as UTF-8.
While Ruby itself does do this pretty well, and most browsers are good
at guessing the file type/encoding based on the contents of the file,
there might be some outliers that rely on metadata that's a bit more
strictly adhered to. This change ensures that sample CSV/JSON files are
delivered to the user as an attachment, and using the correct MIME type,
so that they register as such when downloaded by the browser.
Previously, all imports were showing as "TXT file" types, when they were
really "CSV file" or "JSON file", and at least in Firefox, they were not
downloading when you clicked the sample link. Instead, a new tab would
open (since Firefox thinks it's a text file), and you have to refresh
the page to actually get the browser to download the file.

WORKAREA-77