Import UTF-8 CSVs With BOM Characters
UTF-8 doesn't need a BOM in order to start or end a file, but these
characters can end up in CSVs generated by older software that doesn't
have great support for Unicode. As a result, if a BOM is in the CSV near
`_id` it will cause improper importing of the data held within. To
address this, Workarea now specifies the `bom|` prefix in the `:encoding`
param for `CSV.foreach` by default. This can still be overridden if you
have an ASCII file, and since BOM stripping doesn't really apply,
developers can override the entire encoding string in configuration
if necessary. But this is a sane default for those who use UTF-8 and
happen to be exporting out of older spreadsheet software.

WORKAREA-79