[DOCS] Improve documentation feedback process
Documentation feedback is collected anonymously through a Google form.
This does not allow for communication with the reporter, and it is an
additional communication channel that has to be managed separately.

Use GitHub issues instead of the Google form to collect documentation
feedback. This allows triaging docs issues using the same process used
for the software and allows communication with the reporter (to thank
them, ask questions, provide a status, or help them if the issue is not
actually a documentation problem).

Also update the documentation issue template to broaden its scope and
provide a default assignee.

WORKAREA-64

No changelog