Fix time zone querying for insights and reports
Data was getting stored correctly, but not queried correctly. When
building aggregations for MongoDB, the Mongoid logic to use UTC does not
kick in since it's not going through the Mongoid DSL.

This was the lowest impact fix. Changing how we store reporting_on will
invalidate current data and not allow for changing timezones later.

WORKAREA-135