[DOCS] Fix href and src values
Restoring the old section ID logic fixed the majority of broken links to
sections, but it broke all links to sections that were written _after_
the migration to the current toolchain.

Fix those href values to use the "old" values to make all well, now and
moving foward.

While doing this, also fix href and src values in general. The current
toolchain requires all href and src values to be root-relative (begin
with "/") for them to be portable (continue working if the article in
which they appear is moved).

WORKAREA-63

No changelog