Fix `#casecmp?` in Traffic Referrer Segment Rule
This method call fails with a `TypeError` on any Ruby version lower than
2.5.0, when the method was changed to support `nil` values in its
arguments. To address this, call `#to_s` on all arguments passed into
the `#casecmp?` calls in `Segment::Rules::TrafficReferrer`. This
prevents an out-of-box test failure if you are running tests on Ruby
2.4.x. Additionally, this commit adds a test matrix to the build that
will ensure all code is tested against the latest versions of Ruby
2.4.x and Ruby 2.6.x, and any other Ruby versions that the platform may
support.

WORKAREA-201