Fix Content Block Copying
When copying a page, the content associated with that page was not
being copied properly. This resulted in unexpected sharing of content
blocks for both pages, with changes to one page affecting both in
tandem. To resolve this, remove any other content records that exist
prior to cloning and saving the content record from the copied page.
Additionally, map over all content blocks and ensure they are cloned so
that the block IDs are not shared between documents.

WORKAREA-168