Fix Regular Price Displays In SKUs
The **Regular Price** value in admin was being aliased to the
`#sell_price`, which can sometimes be the sale price. To remedy this,
Workarea now explicitly defines the `Pricing::Sku#regular_price` method
and defines `Admin::PricingSkuViewModel#on_sale?` to represent whether
any of the prices in a SKU are on sale. The SKU is considered "On Sale"
if this is so.

WORKAREA-211