Divide By Units Sold in Average Price Calculation
When calculating the average price for a product in its insights,
Workarea was previously using the amount of orders the product appears
in as a divisor. This will not show the correct average price of a
product unless every order only has a quantity of 1, since it includes
the total price of the item rather than its unit price. To make this
number accurately reflect the average price paid per unit on a product,
Workarea now uses the number of units sold as the divisor when
calculating the average unit price of a product.

WORKAREA-215