Render Setup When Link is Clicked in Segment Workflow
The "Setup" page link sent users to the edit action, which has no
template associated with it. This resulted in an `UnknownFormat` error
appearing in development and a blank page in production. Workarea now
renders the `:setup` action's template when going back to the setup
page so this will render properly with the right information in the
form.

WORKAREA-182