Allow for fields named "type" in CSV import/export
We hit a scenario where someone named a field `type`, so this change
updates the CSV import/export to use `_type` instead of `type` as the
class name for the model. This follows Mongoid attribute conventions.

This is a breaking change. Imports specifying classes will need to
update their fields.