Fix duplicate products in release previews for featured product changes
When featured product changes stack in a release, duplicates will show
when previewing. This is due to the product's Elasticsearch documents
missing changeset IDs for releases scheduled after the release that
document is for. This fixes by indexing those release IDs as well.

Note that this will require a reindex to see the fix immediately. But
there's no harm in letting it roll out as products gradually get
reindexed.