Fix Promo Code Counts in Admin
Previously, promo codes could only be generated once through the admin,
so rendering the count of all promo codes as the count requested to be
generated was working out. However, as CSV imports and API updates became
more widespread, this began to break down as the `#count` field would
have to be updated each time a new set of promo codes were added.
Instead of reading from this pre-defined field on the code list, render
the actual count of promo codes from the database on the code list and
promo codes admin pages.

WORKAREA-199