Raise Error When URL Not Allowed Through Proxy
The proxy in Commerce Cloud will return a 403 response with a special
header when a request has not been allowed through. In this case,
Workarea now raises an error describing the problem and how to go about
solving it. This resolves an issue where previously, errors thrown as a
result of a misconfigured proxy would be cryptic and hard to debug. With
this more descriptive error, developers will be able to see from Sentry
how they can resolve such a scenario.

WORKAREA-226