Configure Sliced Credit Card Attributes
To prevent an unnecessary decoration of the `Workarea::Payment` class,
the attributes sliced out of the Hash given to `Workarea::Payment#set_credit_card`
is now configurable in initializers. This same set of attributes is also
used in the `Users::CreditCardsController`, so the configuration will be
reused when users are attempting to add a new credit card to their
account.

WORKAREA-257