Shipping SKUs Admin UI
This adds an interface to Shipping SKUs from within the admin. Based on
the Pricing SKUs admin UI, it allows for administration of a SKU's
weight and dimensions. This information is used by ActiveShipping in
order to generate a package and automatically obtain shipping rates for
the item represented by this SKU.

WORKAREA-228