Handle `nil` Percentages in Tax Rates UI
The `TaxApplication` module already handles percentages that are not
present, but the tax rates UI expects values there. This results in some
avoidable 500 errors within admin when you blank out a tax rate
percentage field. To resolve this, Workarea now makes sure that all
percentages are of type `Float`, so they can be displayed as "0%" in
the admin whenever a `nil` value is encountered.

WORKAREA-278