Add Multiplier for Bulk Editing Product Pricing
When bulk editing products, Workarea previously allowed the admin to set
pricing on all SKUs in bulk. This was always a flat amount, and all
prices were updated to match that amount. To build upon this feature,
bulk editing products will now enable admins to set a percentage or
amount they wish to increase/decrease the prices by. This allows, for
example, a merchandiser to mark up all products in their bulk edit query
20%, or mark down all products by $15. The existing functionality for
setting these prices has also been preserved, and extended to allow for
setting the price to a percentage of the existing price, such as setting
the price of a $20 to 20%, which would cause the product's price to be
$4.

WORKAREA-276