Add Note To Category Default Sort Edit
The selected `default_sort` of a category will be always used in the
storefront. If the category contains featured products, this sort will
be labelled "Featured", and this might prove confusing to some admins.
To resolve this, add a note just below the dropdown indicating what will
occur when products are featured in the category.

WORKAREA-289